+++ examples/WebClient/WebClient.ino
@@ -47,7 +47,7 @@
 #define TINY_GSM_RX_BUFFER 650
 
 // See all AT commands, if wanted
-//#define DUMP_AT_COMMANDS
+// #define DUMP_AT_COMMANDS
 
 // Define the serial console for debug prints, if needed
 #define TINY_GSM_DEBUG SerialMon
@@ -57,10 +57,10 @@
 #define GSM_AUTOBAUD_MAX 115200
 
 // Add a reception delay - may be needed for a fast processor at a slow baud rate
-//#define TINY_GSM_YIELD() { delay(2); }
+// #define TINY_GSM_YIELD() { delay(2); }
 
 // Uncomment this if you want to use SSL
-//#define USE_SSL
+// #define USE_SSL
 
 // Define how you're planning to connect to the internet
 #define TINY_GSM_USE_GPRS true
@@ -210,7 +210,7 @@ void loop() {
   client.print("Connection: close\r\n\r\n");
   client.println();
 
-  unsigned long timeout = millis();
+  uint32_t timeout = millis();
   while (client.connected() && millis() - timeout < 10000L) {
     // Print available data
     while (client.available()) {
